Add EditableText to support label/hint in material text field
Please note that CL does not fix issue where traversing with granularity and selecting to speak multiple merged texts is possible. For text field it means that although both label and editable text are announced, we will only traverse or select to speak the editable text.
To be able to traverse a merged texts/editable texts, we need to get text layout result for each of them and be able to map text to text layout result. We have a bug b/157474582 to support it that is referenced in the code.

Test: added a few tests, existing tests passed
Bug: 153645139
Relnote: "Added EditableText semantics to mark editable input text of the text field for accessibiity and correponding test methods to check the semantics"
Change-Id: I8e07ab8a8f29d5c2a31fb1c979e099303083a38f