Extract double and long click functionality into the separate exterimental combinedClickable.
Due to the design issue with double click with mouse we have to carefuly reconsider how click and doubleclick cooperates in clickable. Therefore, this CL extracts the double and long click to a separate component and marks is experimental so we will be able to revisit this later on.

Change-Id: Iafad17bbdb5029ce413bce94cd47f7c5233cb67b
Relnote: Modifier.clickable now doesn't have double and long click support. Use Modifier.combinedClickable to achieve this functionality.
Test: adjusted accordingly
Bug: 178495376