Fix capture session configuration error for CameraX test activities
In Camera2TestActivity, the preview's SurfaceTexture takes on the size of its associated TextureView, which is not necessarily a supported camera resolution. A way to work around this is to set the preview's surface to a guaranteed resolution, e.g. 640x480.

In CameraXTestActivity, the preview's (TextureView's) SurfaceTexture is reset to a manually created SurfaceTexture. This seems to cause issues on certain API levels (e.g. 21). Instead, the original SurfaceTexture should be used and resized to the camera's requested resolution.

Bug: 178759621
Test: Manual testing + CameraDisconnectTest.java
Change-Id: I2769527fa53fa74b12b90f7965c7d06e68d33f80