Check lifecycle state before bind use case
The ProcessCameraProvider.getInstance() might be blocked if the previous
CameraX.shutdown task takes a long time.
When the above case happen, we might need to check if the activity is still
in the foreground before start binding the use case.

If we bind the use case when the test activity is already
destroyed, It will throw the exception:
"java.lang.IllegalArgumentException: Trying to create LifecycleCamera
with destroyed lifecycle.".

The change will check the activity status before start binding.

Bug: 178127576
Test: Manually run all tests under uiwidget test app
Change-Id: I51f24542449f02bd59fccc211003f165ab721249