Don't generate synthetic accessors for non-kotlin sources
This CL fixes a bug in KspTypeElement where we would generate synthetic
getters/setters for regular java fields (either in source or .class).

Synthetic getters/setters are necessary only for kotlin properties that
do not specify a getter/setter. For other cases (.java, .class),
there will either be a real accessor (that we'll get from
property.getter/setter) or no need for a synthetic.

Bug: 179075543
Test: XTypeElementTest
Change-Id: I59e31765eab535f411ad15e3d7a2151c18ead229