Add CameraProvider in Core
LifecycleCameraProvider contains the method hasCamera() which isn't related to lifecycles. Core should instead contain a generic CameraProvider which would provide access to a set of cameras, and would contain methods such as hasCamera() and getAvailableCameraInfos().

LifecycleCameraProvider will extend CameraProvider and add lifecycle-related methods.

ProcessCameraProvider will continue to extend LifecycleCameraProvider, with a lifecycle/scope matching that of the application's process.

Relnote: N/A
Bug: 178719290
Test: N/A
Change-Id: I400e5a97474c339e32dba406b00371ee60fbc24f