Account for lambda line numbers from inline functions
If a lambda is calling an inline function, the line numbers of the
inlined instructions are added to the end of the file.
Account for this by analyzing the line number table and ignore
lines that doesn't fall into the first range of lines.

Also the newer compiler 1.4.30 was generating more methods with
line numbers in JVMTI. Select the method with the widest range.

Fixes: 179678965
Fixes: 179040867
Test: Unit test added for an inline lambda
Change-Id: Iff6cf09a54354d7bd36011e0209e8cde9c840e64