Ensure all saved transactions have reordering allowed
The documentation specifically calls out that
setReorderingAllowed(true) should be used on each
FragmentTransaction (see
https://developer.android.com/guide/fragments/transactions#reordering)
but this is moves from a 'should use' to a
'must use' pattern when it comes to saving the
back stack and the FragmentTransactions on it.

This is because setReorderingAllowed(true) is
critical to ensuring that a set of back to back
transactions are executed as an atomic operation,
thus preventing any intermediate fragments (those
that are added and then moved onto the back stack)
from moving through lifecycle states unnecessarily.

By checking this when executing saveBackStack(), we
ensure that the later restore operates as expected.

Test: new saveNonReorderingAllowedTransaction test
BUG: 80029773
Change-Id: Ied3b2a221202ca4626975853b741c76edd08f978