Prevent LegacyPagingSource from passing negative indices
This CL fixes a bug in the legacy path where we were not bounds checking
the parameters we send, possibly causing a negative initial position
that could crash unsuspecting callbacks.
I've updated the PositionalDataSource to assert those values when
initialized so that if we ever calculate them wrong, it crashes
immediately instead of failing silently (there were actually tests
hitting this in PositionDataSourceTest but those values were not
asserted).

Also added a test to LivePagedListTest to make sure if an instant
executor is used, initial paged is computed syncronously as well as any
additional items when they are accessed.

Bug: 180542492
Test: LivePagedListTest, PositionalDataSourceTest
Change-Id: I25b81bd422f7c6720279f06756e860a9515c7349