Add BackStackState infrastructure
A saved back involves two main components:
1. The set of FragmentTransactions that make
up that saved back stack
2. The set of Fragments that were added
as part of that back stack

This is encapsulated in the new BackStackState
class. This is class that will eventually be
created by saveBackStack() and later used by
restoreBackStack() to restore the stack. Those
integration points are coming in follow up CLs:
this is solely the Parcelable used to store this
state.

Test: new BackStackRecordTest test suite
BUG: 80029773
Change-Id: I1658c01fdf7af8941436db96000e91796b823108