Add a quirk for Samsung A3 stretched preview.
The quirk is that the preview is always stretched, and it seems that the FOV is always 1/3 wider than it should be.

The fix is changing the transform in PreviewView so that the extra FOV gets cropped.

This fix is not perfect as it won't fix core users. e.g. core test app still has the issue. However, there isn't a good way to fix it without adding new API to provide the extra info about stretching. With this fix at least the view artifact will work as expected.

Relnote: fix Samsung A3 stretched preview in PreviewView.

Bug: 180121821
Test: manual test and ./gradlew bOS
Change-Id: Iacb30b0ff035d5a24ac221470a6815d96eca1725