Use lazy invalidation tracking for Paging 3
This CL works around a bug in Paging3 integration where using Paging3
with Room could cause main thread queries.

This is a workaround, not a proper fix.

In Paging2, we had a data source factory class and its create method was
always called on a background thread. It also had very strict
invalidation requirements such that if database changed underneath, we
had to detect it before returning any data.

In Paging3, the factory does not exist anymore as a concrete class so a
common implentation is to do:

```
val myPager = Pager(...) {
   myDao.createPagingSource()
}
```

That code is invoked in the collection dispatcher which is commonly the
main dispatcher. Because we are still using the Paging2 APIs internally,
this would end up creating the old data source impl (Paging2) and then
wrap it into a Paging3 PagingSource. Unfortunately, creating Paging2
DataSource requires syncronizing invalidation trackers with the backing
database, hence possibly a main thread query.

The ultimate plan for Paging3 support is to create a room-paging
artifact. This CL is a quick fix for the problem above where we change
the behavior of the LimitOffsetDataSource not to register an
invalidation when it is created, instead, register it when we run the
very first query. This way, it is more lazy (good) and won't try to load
the data on the main thread (Paging3 wrapper for Paging2 data source
takes care of moving load functions to the background thread).

Bug: 181221318
Test: PagingSourceTest
Relnote: "Fixed a bug where creating PagingSource on the main thread
could trigger an ANR."

Change-Id: I42b74e397bcf717f509b1d53203d2edced89afb0