add semantics liveRegion and apply it to snackbar
Snackbar should be marked as live region for accessibility. When talkback
is on, the snackbar is announced when it appears. Also added dismiss
action for the snackbar.

Note that since we are not merging up the tree for subtree change
events, we don't need to do extra work for live region. From android
documentation and discussion with Phil, the throttling delay is fine for
accessibility events from live region.

Relnote: "LiveRegion accessibility API is added.
If node is marked as a live region, the accessibility services
will automatically notify the user about its changes"

Fix: b/153644625, b/172590946, b/182834687

Test: tested manully with talkback on. Unit tests added.

Change-Id: Idcf6f425b12b005e59ad77fe7430e466132ea87c