Let FragmentViewLifecycleOwner implement ViewModelStoreOwner
SavedStateViewModelFactory makes some assumptions that
the SavedStateRegistryOwner is also a ViewModelStoreOwner.
This is almost always the case, except when using the
ViewTree APIs in Fragments. This change makes it so
that the internal FragmentViewLifecycleOwner that is
tied to the Fragment's View Lifecycle implements
ViewModelStoreOwner as well (delegating to the Fragment's
ViewModelStore).

Relnote: "Using `SavedStateViewModelFactory` now works
when used with the `SavedStateRegistryOwner` returned by
using `findViewTreeSavedStateRegistryOwner()` from a
Fragment's View."
Test: updated tests pass
BUG: 181577191

Change-Id: I21acf18f70ab6dcc60f5946d5b0d878e12bc76df