Introduce new artifact emoji2:emoji2-views
This contains the View related classes that previously lived in
emoji:emoji. These classes are not necessary when using EmojiCompat with
appcompat and should only be included for apps that don't use appcompat.

Helpers have been moved to a core artifact emoji2:emoji2-views-core to
allow them to be included from appcompat (but not included by Compose)

Bug: b/178035684
Test: Existing tests, manual testing, some test refinements
Change-Id: I90966c49ece727aa3299adfd771cffc00fccba4c