Fix timing test app launch crash on the Samsung Galaxy S20+
Root Cause: The specific Camera (CameraID 4) of the S20+ does not support JPEG. The function, getOutputSizes(ImageFormat.JPEG), returns null and make the NullPointerException when accessing it.

The cl adds the error handle and skips to add the problematic Camera in app's Camera list. It is not allowable to use this Camera in app to do performance test.

Bug: 181833945
Test: manual launch on Galaxy S20+ and Pixel 2 for camera-testapp-timing
Change-Id: I32db0a8556612b09de61e9eb369315d2fb7de897