OnReceiveContentListener: Handle IME insertion permissions release
Currently the platform assumes that it can automatically revoke URI
permissions any time after the InputContentInfo instance is garbage
collected and InputContentInfo.releasePermission() has not been called
explicitly. This can cause permissions to be revoked before the app
has had a chance to read the content when using OnReceiveContentListener,
since we did not keep a reference to InputContentInfo.

This change sets the InputContentInfo object into the extras bundle
passed to OnReceiveContentListener. This way a reference is maintained
and permissions will not be revoked prematurely (apps that were using
a workaround to cache InputContentInfo instances can cache ContentInfo
instances instead when using OnReceiveContentListener).

Bug: 181365512
Test: ./gradlew :appcompat:appcompat:connectedCheck --info --daemon -Pandroid.testInstrumentationRunnerArguments.class=androidx.appcompat.widget.AppCompatEditTextReceiveContentTest
Test: Manually
Change-Id: Icf456a052488d6d7b32cb59a1deac86a38454591