Handle the onConfigureFailed on RELEASED state
The CameraCaptureSession configure fail might have different reaction
(1) Throw the exception when calling the CameraDevice#createCaptureSession
(2) Call the CameraCaptureSession.StateCallback#onConfigureFailed
(3) Throw the exception when calling the
CameraDevice#createCaptureSession, and then call the CameraCaptureSession.StateCallback#onConfigureFailed

The CL is going to fix for case 3. Handle the onConfigureFailed in released
state. In this case, the onConfigureFailed callback is called after the
exception, the state of the CaptureSession is already changed to
RELEASED state, this case should not throw an error.

Test: manually run the CaptureSessionTest on lab devices.
Change-Id: Ibdc736518749bd936ea81c20ff0da31552b48b03