Remove node is Ready assertions from drawing
Previously I added checks that the LayoutNode is in Ready state(which means no measure/layout were scheduled) when it is about to be drawn. This check helped to stabilize our measuring logic back then when we had a lot of corner cases not covered, like we were able to accidentally forget to remeasure some nodes. This logic was refactored, stabilized and covered with tests and now these assertions are not really needed as now they make some valid cornercases crashing. Example is:

val state = mutableStateOf(0)
val derived = derivedStateOf { 1 }
Canvas(Modifier.size(20.dp)) {
    derived.value // during draw
}
Layout({}) { _, _ ->
    layout(10, 10) {
         state.value = state.value + 1 // read and write during layout
    }
}

Here during the layout invocation the state is read and written, this is not a good pattern as it introduces backwards write, but we can't really completely disallow it. Then during the drawing we first draw a Canvas. Getting the value from the derived state is doing snapshot.notifyObjectsInitialized() which synchronously calls requestRelayout() on the second Layout. Later when we start drawing the second Layout it crashes as the state is not Ready.
So I think there is no reasons to crash as we can just draw the current state safely.

Fixes: 182821580
Fixes: 178508735
Test: manually, existing tests
Change-Id: Ib295495715f77e467764f7617c06ba32a677b977