Check that layout node is attached before reading its coordinates
From crash log from the bug attached it looks like we traverse the semantics node for corresponding layout nodes that are not attached yet. Changed the positionInRoot and similar methods
to check if the nodes are attached. Also ignore the nodes that are not attached when calculating uncovered semantics nodes.

Fixes: 182551468
Test: new tests in AndroidComposeViewAccessibilityDelegateCompatTest
Change-Id: Id25efd4e5cac8352ef7d80df291735e2ac2b1aa3