Ensure ActivityResult works correctly with custom parcelables
When using launching an activity for result, if the app leaves its process
and uses custom parcelables in the intent returned from the
ActivityResultContract, there will be a ClassNotFoundException.

We need to check the returned intent for extras and if there are any
present we must call setExtrasClassLoader() with the ClassLoader from
the ComponentActivity.

RelNote: "ComponentActivity now properly launches
ActivityResultContracts that create intents with parcelable
extras even if the process is left"
Test: ComponentActivityResultTest
Bug: 182906230

Change-Id: Ieff0578bfd0875f22d42b19040ee24999bb23c89