Make saveBackStack() actually save the back stack state
The saveBackStack() API is responsible for a number of
things:
1) Popping the saved FragmentTransactions off the back
stack - this was done in previous CLs.
2) Saving the actual FragmentTransactions themselves
so that they can later to restored
3) Save the state of each Fragment included in
those FragmentTransactions

This change seeks to address the core points of #2
and #3. Now, when saveBackStack() is called, all
Fragments will go through onSaveInstanceState(), either
immediately after onStop() (mirroring the behavior of
the activity being torn down due to a configuration
change, etc.) or immediately before onDestroy() (in
cases where the Fragment is already stopped - such as
when it is on the back stack).

The FragmentTransactions themselves are also
collapsed (i.e., replacing the expanded set of
REMOVE+ADD operations with the original REPLACE operation)
and have their state saved, readying them for later
restoration.

Note that this change does *not* handle nonconfig state
(ViewModels). That will be done in a follow up CL.

Test: saveBackStack test passes
BUG: 80029773
Change-Id: I37327ab6a7deab0d72d5be1a251c0d5e3a280362