Add support for devices with a chin.
Some round Wear OS devices have a chin, a part of the screen that is
not available for drawing.

This commit adds a way for the watch face to know the size of the chin
and adjust its drawing based on that.

Test: ./gradlew :wear:wear-watchface:test
Relnote: Add chin size to the properties of the device available to the watch face.
Bug: 183899196
Change-Id: I76e1e578f0f32372ba8e540ac0b158b88ff92ba0