Add the API to restore a saved back stack
Saving the back stack is only half of the API;
the other half is the ability to later restore
that back stack onto the FragmentManager. This
is the role of the restoreBackStack() API. It
takes a name (the same name passed to
addToBackStack() and then later passed to
saveBackStack()) and re-instantiates all of the
FragmentTransaction and Fragment instances and
re-applies them to the FragmentManager.

Particularly, this builds on the state saving
done by saveBackStack() to ensure that every
fragment that was saved has its saved state
restored when you call restoreBackStack().

The addition of this API did expose some issues
in the previously added code, namely:
- saveBackStack() needs a callback immediately
after the FragmentTransaction pop is executed and
*before* the Fragments are destroyed. This ensures
that the right internal IDs of Fragments are saved
in the BackStackRecordState and means we can't use
the on commit runnables list and need to maintain
our own list.
- As the population of the BackStackRecordState
is asynchronous, BackStackState needs to keep a
reference to the original list created by
saveBackStack() rather than copying the entries
into its own ArrayList.

Relnote: N/A
Test: new tests pass
BUG: 80029773
Change-Id: Ib8c35cea758379374cde7c1253922e345544b3c5