Handle non-repeated repeatable annotations
This CL fixes a bug in XAnnotationBox where it failed to read Repeatable
annotations when the annotation is not repeated on the element (e.g.
only 1 instance).

Apparentlyy, javac wraps it into the container annotation only if the
annotation is repeated and otherwise, we need to query with the repeated
annotation itself to get the value.

Bug: 160322705
Test: XAnnotationBoxTest
Change-Id: I50cdf4b0ebeeaa6ca9d5e18b80d2aeb6e1757dff