Drop compositions that is not owned the current ComposeView
When e.g. an AlertDialog is shown in a compose application, the
contents will be shown in a separate window. The SlotTree with the
dialog contents will actually show up twice:
- once in the dialog window
- once in the main application

The layout nodes found in the content will however belong to the
dialog window. In this change we will start filtering out the
composition that is present in the SlotTree of the main app by
filtering of the ownerView of the layout nodes.

Fixes: 177080852
Test: Added unit test: testFilterOutAlertDialogFromApp
Change-Id: Iecbde104a785e817ae8d591314a3b175d68ac71f