Fix hide() transitions for new state manager
When doing a hide operation with transitions, the view is gone before
the transition ever gets a chance to start, meaning the transition never
gets seen. The reason for this is that immediately after calling start
transition, we synchronously call applyContainerChanges and move the
view to the GONE state. This means by time we get to the asynchronous
onTransitionStart() the view is already GONE.

If we are doing a hide, we should skip the call to
applyContainerChanges and allow the transition to set the fragment view
to GONE upon completion.

RelNote: "The new fragment state manager now properly handles exit
transitions with hide operations."
Test: FragmentTransitionTest
Bug: 184830265

Change-Id: I9e4de834e866143295435825c6eba8d5b06011f8